Simplify ground fire shader to improve performance
The flame sprite could cause performance problems on some Radeon
hardware. The problems were specifically connected with frequent texture
switches. The flame sprite uses a different texture for each of its 3
stages and the flame sprites are rendered one at a time. Also note that
there are about 14 flame sprites at a time for a single ground fire.

This change reduces the shader to a single stage, so that there will be
no texture switches when consecutively rendering flame sprites.