gfx: introduce firebomb groundfire and use it to replace flamer groundfire
That ground fire is named gfx/weapons/firebomb/fire because its purpose
to be used as firebomb ground fire. Flamer ground fire reuse it for now
as a placeholder.

Images are CC By SA Retimer, WrathGames Studio
https://opengameart.org/content/wgstudio-fire-animation-loop

This is the 256 pixel large (the higher provided), and 25 frames version
as Dæmon engine is currently limited to 24 frames per animation, but there
exist 50 and 75 frames versions of this animation. It also exists smaller
resolutions of the images (down to 20 pixel) and less frame versions (down
to 13 frames) so it would be possible to do low versions of it if needed.