Nuke orange particle hit mark
In 2015, an automatic depth fade with depth scale based on the
particle's radius was introduced for all particles. This caused some
particles to be invisible in many cases because the alpha goes to
to ~0 with the depth fade.

One of these is an orange glowing 'mark' on rifle, shotgun,
and chaingun impact. It is invisible unless you shoot a translucent
surface, or the line of sight from the viewer to the impact site is
nearly tangent to the surface. Seems unnecessary and ugly to me; get rid
of it.