NUKE lasgun impact 'mark' particle
The lasgun impact produces two kinds of particle with the
gfx/weapons/lgun/sparkN shaders: (a) a larger, stationary mark and (b)
scattering smaller particles. This commit gets rid of (a).
As of 0.54, (a) is not visible when viewed head on to the impact
surface. When automatic depth fade is disabled, it will have the same
problems as detailed in
https://github.com/UnvanquishedAssets/res-weapons_src.dpkdir/pull/19.
Plus the sprite is not very good: the edges of the square are clearly
visible (including on the smaller particles).