NUKE rifle & shotgun delayed impact smoke
The rifle and shotgun impact particle systems have a little spray of
dust and rocks at first. Then a couple of seconds later, after the dust
is already gone a while, the smoke starts. This looks stupid IMHO.

The delayed smoke particle is invisible in 0.54 due to automatic depth
fade, so nuking it is just maintaining the status quo.