Fix wrong neutral color for lcannon/grenade heat haze
The majority of the heat haze image used for lcannon and grenade had a
solid color which was presumably intended to produce zero distortion in
those areas. For no distortion the color should be (0.5, 0.5, 1). But
this background color was a bit off, which caused the square edges of
the heat haze texture to be clearly visible in game. I adjusted the
background color using the paint bucket tool in paint.net.