Adjust lcannon primary flare sprite
This particle, which was invisible due to bugs as of 0.54, suffers from
hard edge clipping. It's far too big and close to the surface for depth
fade to help. So just make it have a shorter lifetime. We get a quick
explosion flash and then it disappears before the viewer can make out
its shape in too much detail.