Add depth fade to particles from blaster impact
Set a depth fade for the particles that spring out from the blaster
impact site. Some of the particles pick a direction where they are stuck
on the surface instead of flying out, and the sprite often clips the
surface in that case. With the depth fade the stuck particles will be
mostly invisible intsead.