Make grenade 'flare' smaller, shorter
This particle was not visible in 0.54 due to a culling issue.

Splitting the grenade and lcannon flares into 2 shaders somehow made the
grenade flare start rendering over the smoke instead of under it. Not
sure if this is because the particles are z fighting or if it has to do
with the engine's bug with sorting translucent shaders.