Get rid of grenade/firebomb heat haze
This heat haze was not visible in 0.54 due to shader sorting issues. It
is way too big (and therefore badly clips the ceiling), but if I make it
a reasonable size you can't really see it through the smoke (grenade)
or flames (firebomb) anyway.

Moved grenade dynamic light to the flare particle. That makes it shorter
in duration but that seems good anyway.